requirements.txt
# To create notifications
pyler

# To create the interface
PyQt5
 
# To connect to mysql
mysqlclient
mysql-connector-python

# To play music
pafy
python-vlc
youtube-dl==2020.12.2

# To test code quality
bandit
cohesion
flake8
flake8-docstrings
flake8-polyfill
#metrics.pylint
pylint
radon

# Code style
black

# Documentation
pdoc3
sphinx

# Unit test and coverage
coverage
